backend/requirements.txt
Django~=4.1.7
python-dotenv==1.0.0
djangorestframework~=3.14.0
django-filter~=22.1
Pillow==9.4.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
django-extensions==3.2.1
boto3==1.26.99
django-storages==1.13.2
gunicorn==20.1.0
pytest==7.2.2
pytest-django==4.5.2
